Start transitioning code to MOOSE style

+++ material.C
@@ -38,7 +38,7 @@ void materialBase::prepare()
   for (unsigned int i = 0; i < element.size(); ++i)
   {
     am += element[i]->m * element[i]->t;
-    az += double( element[i]->z ) * element[i]->t;
+    az += double(element[i]->z) * element[i]->t;
   }
 
   arho = rho * 0.6022 / am; //[TRI00310] atoms/Ang^3
@@ -50,26 +50,26 @@ void materialBase::average(const ionBase * pka)
   mu = pka->m1 / am;
 
   // universal or firsov screening length
-  a = .5292 * .8853 / ( std::pow( double(pka->z1), 0.23 ) + std::pow( az, 0.23 ) );
-  //a = .5292 * .8853 / pow( pow( double(pka.z1), 0.5 ) + pow( az, 0.5 ), 2.0/3.0 );
+  a = .5292 * .8853 / (std::pow(double(pka->z1), 0.23) + std::pow(az, 0.23));
+  //a = .5292 * .8853 / std::pow(pow(double(pka.z1), 0.5) + std::pow(az, 0.5), 2.0/3.0);
 
   // mean flight path0
-  f = a * am / ( az * double(pka->z1) * 14.4 * ( pka->m1 + am ) );
+  f = a * am / (az * double(pka->z1) * 14.4 * (pka->m1 + am));
   //eps0 = e0 * f;
-  epsdg = simconf->tmin * f * std::pow( 1.0 + mu, 2.0 ) / ( 4.0 * mu );
+  epsdg = simconf->tmin * f * std::pow(1.0 + mu, 2.0) / (4.0 * mu);
 
   // fd and kd determine how much recoil energy goes into el. loss and vaccancies
-  fd = std::pow( 0.01 * az, -7.0 / 3.0 );
-  kd = std::pow( 0.1334 * az, 2.0 / 3.0 ) / std::sqrt( am );
+  fd = std::pow(0.01 * az, -7.0 / 3.0);
+  kd = std::pow(0.1334 * az, 2.0 / 3.0) / std::sqrt(am);
 
   for (unsigned int i = 0; i < element.size(); ++i)
   {
     element[i]->my = pka->m1 / element[i]->m;
-    element[i]->ec = 4.0 * element[i]->my / std::pow( 1.0 + element[i]->my, 2.0 );
-    element[i]->ai = .5292 * .8853 / (std::pow( double(pka->z1), 0.23) + std::pow(element[i]->z, 0.23 ));
-    //ai = .5292 * .8853 / pow( pow( double(pka.z1), 0.5 ) + pow( element[i].z, 0.5 ), 2.0/3.0 );
+    element[i]->ec = 4.0 * element[i]->my / std::pow(1.0 + element[i]->my, 2.0);
+    element[i]->ai = .5292 * .8853 / (std::pow(double(pka->z1), 0.23) + std::pow(element[i]->z, 0.23));
+    //ai = .5292 * .8853 / std::pow(pow(double(pka.z1), 0.5) + std::pow(element[i].z, 0.5), 2.0/3.0);
     element[i]->fi = element[i]->ai * element[i]->m /
-                     (double(pka->z1) * double(element[i]->z) * 14.4 * ( pka->m1 + element[i]->m ));
+                     (double(pka->z1) * double(element[i]->z) * 14.4 * (pka->m1 + element[i]->m));
   }
 
   dirty = false;
@@ -91,12 +91,12 @@ double materialBase::rpstop(int z2p, double e)
   int z2 = z2p-1;
   // velocity proportional stopping below pe0
   pe0 = 25.0;
-  pe = std::max( pe0, e );
+  pe = std::max(pe0, e);
 
   // pcoef indices are one less than in the fortran version!
-  sl = (simconf->pcoef[z2][0] * std::pow( pe, simconf->pcoef[z2][1])) +
-       (simconf->pcoef[z2][2] * std::pow( pe, simconf->pcoef[z2][3]));
-  sh = simconf->pcoef[z2][4] / std::pow( pe, simconf->pcoef[z2][5]) *
+  sl = (simconf->pcoef[z2][0] * std::pow(pe, simconf->pcoef[z2][1])) +
+       (simconf->pcoef[z2][2] * std::pow(pe, simconf->pcoef[z2][3]));
+  sh = simconf->pcoef[z2][4] / std::pow(pe, simconf->pcoef[z2][5]) *
        std::log(simconf->pcoef[z2][6] / pe + simconf->pcoef[z2][7] * pe);
   sp = sl * sh / (sl + sh);
   if (e <= pe0)
@@ -106,7 +106,7 @@ double materialBase::rpstop(int z2p, double e)
       velpwr = 0.25;
     else
       velpwr = 0.45;
-    sp *= std::pow( e/pe0, velpwr );
+    sp *= std::pow(e/pe0, velpwr);
   }
   return sp;
 }
@@ -127,7 +127,7 @@ double materialBase::rstop(const ionBase * ion, int z2)
   double vfermi = simconf->scoef[z2-1].vfermi;
   //double atrho = simconf->scoef[z2-1].atrho;
 
-  if( ion->m1 == 0.0 )
+  if (ion->m1 == 0.0)
     m1 = mm1;
   else
     m1 = ion->m1;
@@ -158,17 +158,17 @@ double materialBase::rstop(const ionBase * ion, int z2)
     vrmin = 1.0;
     v = std::sqrt(e / 25.0) / vfermi;
 
-    if( v >= 1.0 )
-      vr = v * vfermi * (1.0 + 1.0 / ( 5.0 * v*v));
+    if (v >= 1.0)
+      vr = v * vfermi * (1.0 + 1.0 / (5.0 * v*v));
     else
       vr = (3.0 * vfermi / 4.0) * (1.0 + (2.0 * v*v / 3.0) - std::pow(v, 4.0) / 15.0);
 
     yr = std::max(yrmin, vr / std::pow(fz1,0.6667));
     yr = std::max(yr, vrmin / std::pow(fz1,0.6667));
-    a = -0.803 * std::pow( yr, 0.3 ) + 1.3167 * std::pow( yr, 0.6 ) + 0.38157 * yr +  0.008983 * yr*yr;
+    a = -0.803 * std::pow(yr, 0.3) + 1.3167 * std::pow(yr, 0.6) + 0.38157 * yr +  0.008983 * yr*yr;
 
     // ionization level of the ion at velocity yr
-    q = std::min(1.0, std::max( 0.0, 1.0 - std::exp(-std::min(a, 50.0))));
+    q = std::min(1.0, std::max(0.0, 1.0 - std::exp(-std::min(a, 50.0))));
 
     b = (std::min(0.43, std::max(0.32, 0.12 + 0.025 * fz1))) / std::pow(fz1, 0.3333);
     l0 = (0.8 - q * std::min(1.2, 0.6 + fz1 / 30.0)) / std::pow(fz1, 0.3333);
@@ -179,25 +179,25 @@ double materialBase::rstop(const ionBase * ion, int z2)
       q1 = 0.2;
       l1 = b * (q - 0.2) / std::abs(std::max(0.0, 0.9 - 0.025 * fz1) - 0.2000001);
     }
-    else if(q < std::max(0.0, 1.0 - 0.025 * std::min(16.0, fz1)))
+    else if (q < std::max(0.0, 1.0 - 0.025 * std::min(16.0, fz1)))
       l1 = b;
     else
-      l1 = b * ( 1.0 - q ) / ( 0.025 * std::min( 16.0, fz1 ) );
+      l1 = b * (1.0 - q) / (0.025 * std::min(16.0, fz1));
 
-    l = std::max( l1, l0 * lfctr );
-    zeta = q + ( 1.0 / ( 2.0 * vfermi*vfermi ) ) * ( 1.0 - q ) * std::log( 1.0 + sqr( 4.0 * l * vfermi / 1.919  ) );
+    l = std::max(l1, l0 * lfctr);
+    zeta = q + (1.0 / (2.0 * vfermi*vfermi)) * (1.0 - q) * std::log(1.0 + sqr(4.0 * l * vfermi / 1.919 ));
 
     // add z1^3 effect
-    a = -sqr( 7.6 - std::max( 0.0, std::log( e ) ) );
-    zeta *= 1.0 + ( 1.0 / (fz1*fz1) ) * ( 0.18 + 0.0015 * fz2 ) * expf( a );
+    a = -sqr(7.6 - std::max(0.0, std::log(e)));
+    zeta *= 1.0 + (1.0 / (fz1*fz1)) * (0.18 + 0.0015 * fz2) * std::exp(a);
 
     if (yr <= std::max(yrmin, vrmin / std::pow(fz1, 0.6667)))
     {
       // calculate velocity stopping for  yr < yrmin
-      vrmin = std::max( vrmin, yrmin * std::pow( fz1, 0.6667 ) );
-      vmin = 0.5 * ( vrmin + std::sqrt( std::max( 0.0, vrmin*vrmin - 0.8 * vfermi*vfermi ) ) );
+      vrmin = std::max(vrmin, yrmin * std::pow(fz1, 0.6667));
+      vmin = 0.5 * (vrmin + std::sqrt(std::max(0.0, vrmin*vrmin - 0.8 * vfermi*vfermi)));
       eee = 25.0 * vmin*vmin;
-      sp = rpstop( z2, eee );
+      sp = rpstop(z2, eee);
 
       if (z2 == 6 || ((z2 == 14 || z2 == 32) && z1 <= 19))
         power = 0.375;
